update cart add and view

+++ ecommerce-backend/src/main/java/com/example/ecommerce_backend/dto/response/product/ProductResponse.java
@@ -2,6 +2,7 @@ package com.example.ecommerce_backend.dto.response.product;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 @Getter
 @Setter
@@ -15,8 +16,8 @@ public class ProductResponse {
 
     private String thumbnail;   // ⭐ ẢNH ĐẠI DIỆN
 
-    private Double minPrice;    // ⭐ Giá thấp nhất
-    private Double maxPrice;    // (optional)
+    private BigDecimal minPrice;    // ⭐ Giá thấp nhất
+    private BigDecimal maxPrice;    // (optional)
 
     private List<String> categories;
 }
